outsourced FixHoverBehavior and ImportDockingFix
