+++ app/lib/objectivemodeler/features/modeling/behavior/index.js
@@ -1,7 +1,7 @@
 import AdaptiveLabelPositioningBehavior from './AdaptiveLabelPositioningBehavior';
 import AppendBehavior from './AppendBehavior';
-import FixHoverBehavior from './FixHoverBehavior';
-import ImportDockingFix from './ImportDockingFix';
+import FixHoverBehavior from '../../../../frequentlyUsedFiles/features/modeling/behavior/FixHoverBehavior';
+import ImportDockingFix from '../../../../frequentlyUsedFiles/features/modeling/behavior/ImportDockingFix';
 import LabelBehavior from './LabelBehavior';
 import UnclaimIdBehavior from './UnclaimIdBehavior';
 import EmptyTextBoxBehavior from './EmptyTextBoxBehavior';
